+++ release_notes.md
@@ -4,6 +4,10 @@
 
 - Changed default rig environment name from iblenv to iblrig
 - Removed ibllib dependecies
+- Added support for F2TTLv2
+- params.update_params_file() now has a default behavior with no data to update all updatable params
+- removed unnecessary file task_settings.py for calibration
+- SPH for f2ttl calibration now will look for user_settings and fallback on fake_user_settings
 
 ## **Release Notes 6.5.3**